# hashview/api_docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Hashview API
  version: 0.8.2
  description: >-
    The `/v1` API shared by Hashview's human consumers (scripting against the
    server with a personal API key) and the distributed cracking agents.


    **Authentication.** A single cookie named `uuid` carries the credential.
    Depending on the route it must resolve to either a **user's** API key
    (`Profile -> Generate API Key`) or a registered **agent's** UUID — the
    per-operation security requirement (`userApiKey`, `agentUuid`, or both)
    tells you which. On an auth failure the server answers **HTTP 302** to
    `/v1/not_authorized`; following that redirect yields HTTP 200 with an
    error-shaped body.


    **Response envelope.** Almost every endpoint answers **HTTP 200** and puts
    the real outcome in the body's `status` field (200, 400, 403, 426 or 500)
    — see the `ApiMessage` / `ApiError` schemas. Only a handful of newer
    endpoints also use genuine HTTP status codes (404s on
    `DELETE /v1/jobs/{job_id}`, `GET /v1/rules/{rules_id}` and
    `GET /v1/wordlists/{wordlist_id}`); these are modeled explicitly.


    **Serialized rows.** List/detail endpoints return ORM rows JSON-encoded
    **as a string** (double-decode required; see the per-field descriptions).
    The encoder also emits SQLAlchemy artifacts (`query`, `query_class`,
    `registry`, relationship keys such as `owner`) as `null`, and **all
...
    array with ISO-8601 timestamps.
servers:
  - url: /

...
security:
  - userApiKey: []
  - agentUuid: []
...
paths:
...
  /v1/tasks/add:
    post:
      tags: [Tasks]
      summary: Create a task (wordlist + optional rule)
      description: >-
        Creates a hashcat attack-mode 0 (straight dictionary) task. `rule_id`
        is optional — omit it (or send `null`/`"None"`/`""`) for a plain
        dictionary attack. Task names must be unique.
      operationId: addTask
      x-audience: user
      security:
        - userApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            examples:
              withRule: {value: {name: rockyou-best64, wl_id: 1, rule_id: 2}}
              plainDictionary: {value: {name: rockyou-only, wl_id: 1}}
      responses:
        '200':
          description: >-
            HTTP 200 always; body status 200 with `task_id` on success, 400
            for each validation failure, 403 if the key resolves to no user,
            500 on a database error.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TaskAddedResponse'
                  - $ref: '#/components/schemas/ApiError'
              examples:
                added: {value: {status: 200, type: message, msg: Task added, task_id: 42}}
                missingBody: {value: {status: 400, type: Error, msg: Missing task data in request body}}
                missingName: {value: {status: 400, type: Error, msg: Task name is required}}
                duplicateName: {value: {status: 400, type: Error, msg: A task with that name already exists}}
                badWlId: {value: {status: 400, type: Error, msg: wl_id is required and must be a wordlist id}}
                unknownWlId: {value: {status: 400, type: Error, msg: Invalid wl_id}}
                badRuleId: {value: {status: 400, type: Error, msg: rule_id must be a rule id}}
                unknownRuleId: {value: {status: 400, type: Error, msg: Invalid rule_id}}
        '302':
          $ref: '#/components/responses/NotAuthorizedRedirect'
...
components:
  securitySchemes:
    userApiKey:
      type: apiKey
      in: cookie
      name: uuid
      description: >-
        The `uuid` cookie set to a USER's API key (generate one via
        Profile -> Generate API Key). Accepted by operations whose security
        lists `userApiKey`.
    agentUuid:
      type: apiKey
      in: cookie
      name: uuid
      description: >-
        The `uuid` cookie set to a registered AGENT's UUID (agent must be in
        status Online/Working/Idle/Authorized). Accepted by operations whose
        security lists `agentUuid`.
  responses:
    NotAuthorizedRedirect:
      description: >-
        Missing/invalid `uuid` cookie, or a credential of the wrong kind for
        this operation (e.g. an agent UUID on a user-only endpoint) — HTTP
        302 redirect to `/v1/not_authorized`. NOTE: following the redirect
        yields HTTP 200 with an error-shaped body.
      headers:
        Location:
          schema: {type: string, example: /v1/not_authorized}
  schemas:
...
    ApiMessage:
      type: object
      description: >-
        Standard success envelope. The API almost always answers HTTP 200 and
        reports the real outcome in `status` — 200 OK, 400 bad input, 403
        forbidden/unknown user, 426 version mismatch, 500 server error.
      required: [status]
      properties:
        status:
          type: integer
          description: Logical status code (NOT necessarily the HTTP status).
          example: 200
        type:
          type: string
          enum: [message]
        msg:
          type: string
    ApiError:
      type: object
      description: >-
        Standard error envelope, usually still delivered over HTTP 200
        (except the real 404s noted per-operation). `type` is usually
        "Error", but a few error paths use "message" (e.g. jobtask/status
        failures, hashes/import "Unsupported Hashtype").
      required: [status, msg]
      properties:
        status:
          type: integer
          enum: [400, 403, 404, 426, 500]
        type:
          type: string
          enum: [Error, message]
        msg:
          type: string
...
    TaskCreate:
      type: object
      required: [name, wl_id]
      properties:
        name:
          type: string
          description: Must be unique across tasks.
        wl_id:
          type: integer
          description: Wordlist id (must exist).
        rule_id:
          type: integer
          nullable: true
          description: >-
            Optional rule id (must exist when given). Omit, or send
            null/"None"/"", for a plain dictionary attack.
...
    TaskAddedResponse:
      allOf:
        - $ref: '#/components/schemas/ApiMessage'
        - type: object
          required: [task_id]
          properties:
            task_id: {type: integer}
...